docs: add claude.md (#32)
