+++ crates/y-sweet/src/lib.rs
@@ -5,6 +5,7 @@ pub mod convert;
 pub mod server;
 pub mod server_ext;
 pub mod stores;
+pub mod tracing_setup;
 
 #[cfg(test)]
 mod tests;
