+++ crates/y-sweet/src/main.rs
@@ -1,11 +1,6 @@
 use anyhow::Context;
 use anyhow::Result;
 use clap::{Parser, Subcommand};
-use ddtrace::{
-    formatter::DatadogFormatter,
-    set_global_propagator,
-    tracer::{self, ProviderGuard},
-};
 use serde_json::json;
 use std::{
     env,
@@ -20,6 +15,7 @@ use tracing_subscriber::EnvFilter;
 use url::Url;
 use y_sweet::cli::{print_auth_message, print_server_url};
 use y_sweet::stores::filesystem::FileSystemStore;
+use y_sweet::tracing_setup::init_tracing;
 use y_sweet_core::{
     auth::Authenticator,
     store::{
@@ -161,90 +157,6 @@ async fn get_store_from_opts(store_path: &str) -> Result<Box<dyn Store>> {
         Ok(Box::new(store))
     } else {
         Ok(Box::new(FileSystemStore::new(PathBuf::from(store_path))?))
-    }
-}
-
-fn init_tracing(filter: EnvFilter) -> Result<Option<ProviderGuard>> {
-    use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
-
-    let tracing_disabled = std::env::var("DD_TRACE_ENABLED")
-        .map(|value| matches!(value.as_str(), "0") || value.eq_ignore_ascii_case("false"))
-        .unwrap_or(false);
-
-    if tracing_disabled {
-        let fmt_layer = tracing_subscriber::fmt::layer()
-            .json()
-            .with_timer(tracing_subscriber::fmt::time::UtcTime::rfc_3339())
-            .with_target(false)
-            .with_thread_ids(false)
-            .with_thread_names(false)
-            .with_file(false)
-            .with_line_number(false)
-            .with_current_span(true)
-            .with_span_list(false)
-            .with_span_events(tracing_subscriber::fmt::format::FmtSpan::NONE);
-
-        tracing_subscriber::registry()
-            .with(filter)
-            .with(fmt_layer)
-            .init();
-
-        return Ok(None);
-    }
-
-    set_global_propagator();
-
-    if std::env::var("DD_VERSION").is_err() {
-        std::env::set_var("DD_VERSION", VERSION);
-    }
-
-    let service_name = std::env::var("DD_SERVICE").unwrap_or_else(|_| "y-sweet".to_string());
-
-    match tracer::build_layer(service_name) {
-        Ok((datadog_layer, guard)) => {
-            let fmt_layer = tracing_subscriber::fmt::layer()
-                .json()
-                .with_timer(tracing_subscriber::fmt::time::UtcTime::rfc_3339())
-                .with_target(false)
-                .with_thread_ids(false)
-                .with_thread_names(false)
-                .with_file(false)
-                .with_line_number(false)
-                .with_current_span(true)
-                .with_span_list(false)
-                .with_span_events(tracing_subscriber::fmt::format::FmtSpan::NONE)
-                .event_format(DatadogFormatter);
-
-            tracing_subscriber::registry()
-                .with(filter)
-                .with(fmt_layer)
-                .with(datadog_layer)
-                .init();
-
-            Ok(Some(guard))
-        }
-        Err(err) => {
-            let fmt_layer = tracing_subscriber::fmt::layer()
-                .json()
-                .with_timer(tracing_subscriber::fmt::time::UtcTime::rfc_3339())
-                .with_target(false)
-                .with_thread_ids(false)
-                .with_thread_names(false)
-                .with_file(false)
-                .with_line_number(false)
-                .with_current_span(true)
-                .with_span_list(false)
-                .with_span_events(tracing_subscriber::fmt::format::FmtSpan::NONE);
-
-            tracing_subscriber::registry()
-                .with(filter)
-                .with(fmt_layer)
-                .init();
-
-            eprintln!("datadog tracer initialization failed, continuing without APM: {err}");
-
-            Ok(None)
-        }
     }
 }
